added swagger-at-build-time
Now the Swagger file is produced at build time. I added a `Directory.Build.props` file as we discussed and it worked great. In the props file, there are targets to support everything we need to do to build the Swagger file during publish. This will be useful in describing to engineering an example of how we can achieve this.